make quiet really quiet
"--quiet" was ignored in one spot in fdpr_prof.

"--quiet" was being observed in the rest of the wrapper scripts,
but not being passed to fdpr_dl.py.  Pass it to fdpr_dl.py
and implement it there.

Also, delete some incorrect code in fdpr_dl.py which was
never utilized.  (getopt should be called with sys.argv[1:])

DCO 1.1 Signed-off-by: Paul A. Clarke <pc@us.ibm.com>